Refactor date conversion into a dedicated class.
